backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bitarray==2.9.2
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.54.1
frozenlist==1.4.1
fuzzywuzzy==0.18.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
kiwisolver==1.4.7
llamaapi==0.1.36
MarkupSafe==3.0.2
matplotlib==3.9.2
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.1.0
openai==1.54.4
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-sat==1.8.dev13
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.13.1
